Create a 'useExpandButton' util (in e2e tests):
* Create said function, returning and interface used to interact with the expandable field
* Replace all expandable field (expand button) functionality across e2e helpers to use said interface
